@@ -7,57 +7,58 @@
 	"private": true,
 	"author": "",
 	"license": "ISC",
+	"packageManager": "bun@1.1.26",
 	"scripts": {
-		"test": "vitest run",
 		"tdd": "vitest",
-		"lint": "tsc && eslint .",
+		"test": "vitest run",
+		"lint": "tsc && eslint --flag unstable_ts_config .",
 		"dev": "vite",
 		"build": "rm -rf ./dist && vite build",
 		"preview": "vite preview"
 	},
 	"dependencies": {
 		"@kavsingh/curry-pipe": "0.3.0",
-		"@solidjs/router": "0.14.1",
-		"solid-js": "1.8.19",
-		"tailwind-merge": "2.4.0"
+		"@solidjs/router": "0.14.3",
+		"solid-js": "1.8.22",
+		"tailwind-merge": "2.5.2"
 	},
 	"devDependencies": {
+		"@eslint/js": "9.9.1",
+		"@kavsingh/eslint-plugin-filenames": "2.0.0-rc.0",
 		"@solidjs/testing-library": "0.8.9",
 		"@tailwindcss/container-queries": "0.1.1",
 		"@testing-library/dom": "10.4.0",
-		"@testing-library/jest-dom": "6.4.8",
+		"@testing-library/jest-dom": "6.5.0",
 		"@testing-library/user-event": "14.5.2",
 		"@tsconfig/strictest": "2.0.5",
-		"@types/bun": "1.1.6",
-		"@types/eslint": "8.56.10",
-		"@types/node": "22.1.0",
-		"@typescript-eslint/eslint-plugin": "8.0.1",
-		"@typescript-eslint/parser": "8.0.1",
+		"@types/bun": "1.1.8",
+		"@types/eslint": "9.6.1",
+		"@types/eslint__js": "8.42.3",
 		"@vitest/coverage-v8": "2.0.5",
 		"autoprefixer": "10.4.20",
-		"eslint": "8.57.0",
+		"eslint": "9.9.1",
 		"eslint-config-prettier": "9.1.0",
-		"eslint-import-resolver-typescript": "3.6.1",
-		"eslint-plugin-deprecation": "3.0.0",
-		"eslint-plugin-filenames": "1.3.2",
-		"eslint-plugin-import-x": "3.1.0",
+		"eslint-import-resolver-typescript": "3.6.3",
+		"eslint-plugin-import-x": "4.1.1",
 		"eslint-plugin-jest-dom": "5.4.0",
 		"eslint-plugin-prettier": "5.2.1",
-		"eslint-plugin-solid": "0.14.1",
+		"eslint-plugin-solid": "0.14.3",
 		"eslint-plugin-tailwindcss": "3.17.4",
-		"eslint-plugin-testing-library": "6.2.2",
+		"eslint-plugin-testing-library": "6.3.0",
 		"eslint-plugin-vitest": "0.5.4",
-		"jsdom": "24.1.1",
+		"globals": "15.9.0",
+		"jsdom": "25.0.0",
 		"prettier": "3.3.3",
-		"tailwindcss": "3.4.7",
+		"tailwindcss": "3.4.10",
 		"typescript": "5.5.4",
-		"vite": "5.3.5",
+		"typescript-eslint": "8.4.0",
+		"vite": "5.4.3",
 		"vite-plugin-checker": "0.7.2",
 		"vite-plugin-deadfile": "1.3.0",
 		"vite-plugin-html": "3.2.2",
 		"vite-plugin-singlefile": "2.0.2",
 		"vite-plugin-solid": "2.10.2",
-		"vite-tsconfig-paths": "4.3.2",
+		"vite-tsconfig-paths": "5.0.1",
 		"vitest": "2.0.5",
 		"vitest-canvas-mock": "0.3.3"
 	}
